♻️ refactor: split sandbox env config
