+++ src/envs/app.ts
@@ -33,8 +33,6 @@ const ASSISTANT_INDEX_URL = 'https://registry.npmmirror.com/@lobehub/agents-inde
 
 const PLUGINS_INDEX_URL = 'https://registry.npmmirror.com/@lobehub/plugins-index/v1/files/public';
 
-const emptyStringToUndefined = (value: unknown) => (value === '' ? undefined : value);
-
 export const getAppConfig = () => {
   return createEnv({
     clientPrefix: 'NEXT_PUBLIC_',
@@ -76,19 +74,6 @@ export const getAppConfig = () => {
        * e.g., "lobechat-com", "lobehub-desktop"
        */
       MARKET_TRUSTED_CLIENT_ID: z.string().optional(),
-
-      SANDBOX_PROVIDER: z.preprocess(
-        emptyStringToUndefined,
-        z.enum(['market', 'onlyboxes']).optional(),
-      ),
-      ONLYBOXES_BASE_URL: z.preprocess(emptyStringToUndefined, z.string().url().optional()),
-      ONLYBOXES_JIT_ISSUER: z.preprocess(emptyStringToUndefined, z.string().optional()),
-      ONLYBOXES_JIT_SIGNING_KEY: z.preprocess(emptyStringToUndefined, z.string().optional()),
-      ONLYBOXES_JIT_TTL_SEC: z.preprocess(
-        emptyStringToUndefined,
-        z.coerce.number().int().positive().optional(),
-      ),
-      ONLYBOXES_LEASE_TTL_SEC: z.preprocess(emptyStringToUndefined, z.coerce.number().optional()),
 
       AGENT_GATEWAY_SERVICE_TOKEN: z.string().optional(),
       AGENT_GATEWAY_URL: z.string().url().optional(),
@@ -134,13 +119,6 @@ export const getAppConfig = () => {
 
       MARKET_TRUSTED_CLIENT_SECRET: process.env.MARKET_TRUSTED_CLIENT_SECRET,
       MARKET_TRUSTED_CLIENT_ID: process.env.MARKET_TRUSTED_CLIENT_ID,
-
-      SANDBOX_PROVIDER: process.env.SANDBOX_PROVIDER,
-      ONLYBOXES_BASE_URL: process.env.ONLYBOXES_BASE_URL,
-      ONLYBOXES_JIT_ISSUER: process.env.ONLYBOXES_JIT_ISSUER,
-      ONLYBOXES_JIT_SIGNING_KEY: process.env.ONLYBOXES_JIT_SIGNING_KEY,
-      ONLYBOXES_JIT_TTL_SEC: process.env.ONLYBOXES_JIT_TTL_SEC,
-      ONLYBOXES_LEASE_TTL_SEC: process.env.ONLYBOXES_LEASE_TTL_SEC,
 
       AGENT_GATEWAY_SERVICE_TOKEN: process.env.AGENT_GATEWAY_SERVICE_TOKEN,
       AGENT_GATEWAY_URL: process.env.AGENT_GATEWAY_URL,
