+++ src/server/services/sandbox/__tests__/factory.test.ts
@@ -14,8 +14,8 @@ describe('sandbox service factory', () => {
   });
 
   it('uses the market provider by default', async () => {
-    vi.doMock('@/envs/app', () => ({
-      appEnv: {},
+    vi.doMock('@/envs/sandbox', () => ({
+      sandboxEnv: {},
     }));
 
     const { createSandboxService } = await import('../factory');
@@ -35,6 +35,11 @@ describe('sandbox service factory', () => {
   it('uses the onlyboxes provider when configured', async () => {
     vi.doMock('@/envs/app', () => ({
       appEnv: {
+        APP_URL: 'https://lobehub.example.com',
+      },
+    }));
+    vi.doMock('@/envs/sandbox', () => ({
+      sandboxEnv: {
         ONLYBOXES_BASE_URL: 'https://onlyboxes.example.com',
         ONLYBOXES_JIT_SIGNING_KEY: 'jit-signing-key',
         SANDBOX_PROVIDER: 'onlyboxes',
